commone timezone handling code moved into super IngestionModelScraper class
