+++ cdp_scrapers/scraper_utils.py
@@ -1,4 +1,6 @@
+from datetime import datetime
 import logging
+import pytz
 import re
 from typing import Any, List, Optional
 
@@ -67,7 +69,68 @@ def str_simplified(input_str: str) -> str:
 class IngestionModelScraper:
     """
     Base class for events scrapers providing IngestionModels for cdp-backend pipeline
+
+    Parameters
+    ----------
+    timezone: str
+        The timezone for the target client.
+        i.e. "America/Los_Angeles" or "America/New_York"
+        See https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for canonical
+        timezones.
     """
+
+    def __init__(
+        self,
+        timezone: str,
+    ):
+        self.timezone: pytz.timezone = pytz.timezone(timezone)
+
+    @staticmethod
+    def find_time_zone() -> str:
+        """
+        Return name for a US time zone matching UTC offset calculated from OS clock.
+        """
+        utc_now = pytz.utc.localize(datetime.utcnow())
+        local_now = datetime.now()
+
+        for zone_name in pytz.country_timezones("us"):
+            zone = pytz.timezone(zone_name)
+            # if this is my time zone
+            # utc_now as local time should be VERY close to local_now
+            if (
+                abs(
+                    (
+                        utc_now.astimezone(zone) - zone.localize(local_now)
+                    ).total_seconds()
+                )
+                < 5
+            ):
+                return zone_name
+
+        return None
+
+    def localize_datetime(self, local_time: datetime) -> datetime:
+        """
+        Return input datetime with time zone information.
+        This allows for nonambiguous conversions to other zones including UTC.
+
+        Parameters
+        ----------
+        local_time: datetime
+            The datetime to attached timezone information to.
+
+        Returns
+        -------
+        local_time: datetime
+            The date and time attributes (year, month, day, hour, ...) remain unchanged.
+            tzinfo is now provided.
+        """
+        try:
+            return self.timezone.localize(local_time)
+        except (AttributeError, ValueError):
+            # AttributeError: time_zone or local_time is None
+            # ValueError: local_time is not navie (has time zone info)
+            return local_time
 
     @staticmethod
     def get_required_attrs(model: IngestionModel) -> List[str]:
